+++ src/test/java/co/elastic/logstash/filters/elasticintegration/SmokeTest.java
@@ -27,11 +27,42 @@ import java.util.function.Consumer;
 import java.util.function.Predicate;
 
 import static co.elastic.logstash.filters.elasticintegration.EventMatchers.*;
+import static co.elastic.logstash.filters.elasticintegration.EventProcessor.PIPELINE_MAGIC_NONE;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.*;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
 public class SmokeTest {
+    @Test
+    public void givenAFieldWithUserAgentStringTheCorrespondingProcessorIsAbleToParseIt() {
+        final List<Event> inputEvents = List.of(
+                newEvent(Map.of("webbrowser", "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/109.0.0.0 Safari/537.36"),
+                        Map.of("meta", "ok"))
+        );
+        final List<Event> matchedEvents = new ArrayList<>();
+        final EventProcessorBuilder eventProcessorBuilder = EventProcessor.builder()
+                .setEventPipelineNameResolver((event, exceptionConsumer) -> Optional.of("user-agent-mutate"))
+                .setPipelineConfigurationResolver(new LocalDirectoryPipelineConfigurationResolver(getPreparedPipelinesResourcePath("simple-mutate-pipelines")))
+                .setFilterMatchListener(matchedEvents::add);
+
+        withEventProcessor(eventProcessorBuilder, (eventProcessor) -> {
+            final Collection<Event> outputEvents = eventProcessor.processEvents(inputEvents);
+
+            assertThat("event count is unchanged", outputEvents, hasSize(inputEvents.size()));
+            Event firstEvent = outputEvents.iterator().next();
+
+            assertAll("user_agent is correctly decomposed", () -> {
+                assertThat(firstEvent, includesField("[user_agent][version]").withValue(equalTo("109.0.0.0")));
+                assertThat(firstEvent, includesField("[user_agent][os][name]").withValue(equalTo("Windows")));
+                assertThat(firstEvent, includesField("[user_agent][os][version]").withValue(equalTo("10")));
+                assertThat(firstEvent, includesField("[user_agent][name]").withValue(equalTo("Chrome")));
+                assertThat(firstEvent, includesField("[user_agent][device][name]").withValue(equalTo("Other")));
+            });
+
+            assertThat(firstEvent, is(in(matchedEvents)));
+            assertThat(firstEvent, includesField("[@metadata][target_ingest_pipeline]").withValue(equalTo(PIPELINE_MAGIC_NONE)));
+        });
+    }
 
     @Test
     public void testSinglePipelineMutatingEvents() {